checkpoint: Get userId from session inside of API
This code currently successfully gets the userId on form submit. The userID is necessary because when we update the user's profile with their new info, we need to target that data with the userID.

I added a callback inside of nextauth that gets the session and sets the user.id to the token.sub which is basically the user.id (someone found this on GitHub inside of the next-auth code).
- @ ruohola's answer: https://github.com/nextauthjs/next-auth/discussions/536

There were also issues getting session on the server side, since the methods are different from retrieving on client side.
- @ Securing API Routes Section: http://www.topcoder.com/thrive/articles/implementing-authentication-in-next-js?utm_source=thrive&utm_campaign=thrive-feed&utm_medium=rss-feed

Next steps:
- Use this id to target the db row and update the information with the submitted form data.