feat: Header & Layout on all pages
Original plan was to use layout to apply the header to all pages. This would work if it was static, but the one I designed has dynamic text depending on each page.

To fix this, I added dynamic props to the header component and will import it into each page.

Also added visibility classes to hide the logo on certain pages, or hide menu and campaign name on mobile but make it appear on desktop.

More changes soon. WIP.