feat: Add campaign dashboard, add logic for campaigns db
Restructured the form so that when a new user creates an account, they are asked for their role / job title.
This will be automatically pre-filled into their teams cards when they make a new campaign.
If the user wants to change their role, they can do so freely (not implemented yet).
This just skips one step in the process so they don't have to add it every time.

On the dashboard page, it now looks and functions like the Figma design.
There is a lot of extra component logic that isn't needed if I code the component locally
See: https://nextjs.org/docs/api-reference/next/link

Since I'm using Next.js links, the child of the link is a component which means I need to add React.forwardRef and ton of boiler plate code.
After this commit, I'm planning on deleting the component and coding it locally since it's kind of useless to have extra code for 2 lines (div + p tag)