feat: Add menu
When a user selects a campaign, they're presented with a menu screen that lets them navigate to different links for that campaign.

Back button icon also added if they want to go back, using next/router.
https://nextjs.org/docs/api-reference/next/router#routerback

Next step is to create the moodboard page and map all the possible mooodboards, then adjust the menu to include the "Dashboard" link if they're opening the menu from any other page (allows quick access to dashboard)

Might also have to transfer the 100vh webkit Safari fix from the login pages to this.