feat: Add moodboards page
When the user clicks through campaigns to moodboards, it gets the current url via { asPath }. Querying for the data by using the campaignId as a filter to get all of the relevant moodboards.

asPath: String - The path (including the query) shown in the browser without the configured basePath or locale.

https://nextjs.org/docs/api-reference/next/router