feat: Add moodboarding images! + masonry layout, new images
This was the MVP for the project, happy I got this to work. It means I'm on the right track.

When a user uploads an image, it posts to Amazon S3 and generates an image url, this url is then inputted into the database with the corresponding moodboardId.

One problem I faced was getting the new images and old images to look like it was "one" component, rather than two arrays being mapped separately (which was what's actually happening under the hood)
- Solution: Pass props upwards from the child -> parent component by creating a pull data function. The function is run with the image urls as a param inside of a useEffect so it runs after the urls have been generated.
- This gets sent up to the parent so we can map over them. We use array optional chaining to only map over that array if the data exists, meaning that it should only add new images if the user actually uploads images. If not, it will never run.
- This new array is inside of the same container which maintains the masonry styling.

Resources:
- Masonry CSS: https://w3bits.com/css-masonry/