refactor/fix: Paste on Upload works with multiple images
Refactored code to use for of loops and changed the naming to be similar between upload button and paste to upload.

One problem I faced was the code I was using before only worked for one image and was using .items instead of .files, requiring me to convert it again in the code.

In the StackOverflow link, we use Array.from to create an array of files which mimics the upload button function. It also grabs the clipboardData.files instead of items, so we save time on one conversion.

Resources:
- Getting multiple files for clipboardData: https://stackoverflow.com/questions/67977427/extract-files-from-clipboard-or-paste-local-files