style: Add upload button styling
In the last commit I mentioned that I would be changing the images to pull from state through the database instead of passing it up from the child.
- This ended up failing, my previous solution was better. The problem with pulling from state and using router.replace(router.asPath) was that it required a double refresh to show new images, and also on refresh it bumped the page scroll position back to the top which ruined the seamless user experience of adding images.

Upload button has been restyled from default html to now use a label instead. Had to wrap the contents in a form for it to work.

Resources:
- Remove "no file chosen" on file input: https://stackoverflow.com/a/39051660/19416953