feat: Add paste link to upload
When a user pastes a url on the window, it checks if it's a url.
If so, it will also run a check if it's an image url. This url will then get inputted into the database.

If the pasted contents are a file instead, then run upload function normally.

Resources:
- Check if url is an image: https://stackoverflow.com/a/9714891/19416953