refactor: Remove react-dropzone, add custom js dropzone
react-dropzone doesn't allow for the user to click the elements underneath it, which would interfere with future functions like adding a text block, or clicking on images. I honestly think it wasn't even intended to be fullscreen, but rather a div that was only used for uploading.

Another solution was to listen for drag & drop on the window itself, rather than an empty div that took up the size of the window.
Our code no longer requires react-dropzone, you'll see it commented out.
The functions that handle the upload are triggered onDrop and now allows the user to click anything on the page without issues.

Resources
- react-dropzone, fullscreen support?: https://github.com/react-dropzone/react-dropzone/issues/753#issuecomment-774782919