refactor/fix: Refactor conditional if/else on paste to upload
There was a bug where if I pasted an image file, it would think it was an image url since its string ended in .jpeg.
Then there was another bug where it would think it was a string, since it passed both conditions as text AND an image.

I had to refactor the conditions to be more strict, adding an https:// prefix check on the regex and also defining what each type of method would return:

  image copy
  files: [obj...]
  url? false

  text
  files: []
  url: false

  image url
  files: []
  url? true

I used this to create the new conditional statements and differentiate each type so it knows which function should run.