refactor: Remove setContent state, less strict regex on Upload
When a user would add an image, sometimes the url would have queries after the .jpg and it would think the url is plain text instead of an actual image.

ie. from 032c website:
https://cdn.sanity.io/images/wmq4hf8k/production/b65cac68a2b187e9077d8d9b621467d1e3782c9e-1354x2048.jpg?w=1920&q=75&fit=clip&auto=format

or Twitter which uses ?format=jpeg
https://pbs.twimg.com/media/FXaAyVeWQAECDvE?format=jpg&name=large

This was prominent on platforms like Twitter where there's a lot of cool images, but if the user tried to paste it in, they'd only get text and not the photo.

I modified the original regex to not check if it ENDS in .jpg/.png/etc..., but instead if the url itself CONTAINS .jpg/.png That way there can be queries after the true image url and it wouldn't affect the upload process.

Resources
- Test my modified regex: https://regexr.com/6pjul