feat: Wardrobe add item — react-dropzone integrated with react-image-crop
When a user uploads a new item, it prompts them to add an image.
Previously, there wasn't a connection between react-dropzone (file picker) and react-image-crop.

I replaced the default file picker to use the files from react-dropzone and passed the images into the state used by react-image-crop.

Now the user can drag and drop an image, or click on the box to upload one.

Resources:
- URL.createObjectURL, fix for empty file array: https://stackoverflow.com/a/66985035