checkpoint: Add debouncing to rembg component
The dilemma was that the processing time under the current ec2 computer is too slow, approximately 6 seconds for each image to process.

A solution I had was to track if the user had stopped cropping, then we could run the rembg function while the user was filling out the form. That way, the image would be done processing by the time they were finished with the form.

This reduces the time from 6s to essentially 0s, but there were a lot of risks and edge cases.

What happens if the user adjusts the crop after an image is done processing?
  -> We have to delete the old images that were generated. The solution I came up with was to store these images in an array, and at the very end onSubmit, delete all the images in the array except the most recent one.

What happens if the user submits before the image is done processing?
  -> We have to prevent this behaviour. See next question.

How do you wait until the function is done to submit? Do we prevent the user from submitting, or can we queue the function to run after?
  -> I've tried multiple solutions to queue the function which is ideal. The problem is that we are running the background removal function under a useEffect, and we can't use it under a Promise.all. When testing, I noticed that the function inside would run on load, which defeats the purpose of the promise.
  -> Another solution was to prevent the user from submitting at all if the image was still processing or was empty, which is this current code.

How does the backend look when implementing this functionality?
  -> More code. Not as linear and simple.

Is it worth the risk?
  -> 6s down to potentially 0s is huge. I need to test it more and see how much better it really is. Will come back to this question later.
  -> Possibly uses more computing power for unnecessary renders.

After this commit I'm going to revert the code, test how bad the 6s delay is, and then come back to this version.

Resources:
- React Debouncing: https://www.udemy.com/course/react-the-complete-guide-incl-redux/learn/lecture/25599220#overview