checkpoint: Debouncing no cropped image on rembg
There was an edge case where the user didn't crop the image being uploaded (which is totally valid). The solution was to use an if / else statement to check if they were submitting with only an image and not croppedImageURL

The problem with this debouncing solution is that we essentially split the process now:
- Not only are we tracking for activty on croppedImageURL via useEffect,
- We have to do the same for image.
  -> This leads to more code, and the need to create solutions for multiple edge cases
  -> When the user uploads an image, doesn't touch it, then crops after. We now are generating urls for both image and croppedImageURL.

Solution
- if / else statement within the useEffect and have both image + croppedImageURL as a dependency.
- This forces it to only run one rembg function depending on what the current state is

Edge Cases:
- Did the user upload an image and not crop it?
  -> One url generated
- Did they crop it right away?
  -> One url generated (working correctly after these changes)
- Did they upload, forgot to crop, and generate a cropped image afterwards?
  -> Multiple urls generated.
  -> Add urls to array, delete all images except last one onSubmit (need to implement still)

I'm really considering reverting back to the original solution (no debouncing) to keep things simple and prevents the need to think about edge cases.

Resources:
- Placeholder for select button: https://stackoverflow.com/a/62747032
- Variable scope outside of if/else: https://stackoverflow.com/a/13337565
- Convert file to blob (@bgh solution, arrayBuffer): https://stackoverflow.com/a/59284898
- File reader format for images: https://stackoverflow.com/a/42498790 & https://developer.mozilla.org/en-US/docs/Web/API/FileReader/readAsDataURL
- react-image-crop demo for reference: https://codesandbox.io/s/react-image-crop-demo-with-react-hooks-y831o?file=/src/App.tsx

----

next day, update:
I continued to work with the code and I retract my last comment. I think this is the most optimal solution for the user, and I think going forward there needs to be a principle I go by in order to guide my decisions, one of them being: user first.

Over time, this solution will change to become the previous one once we can get servers fast enough to process images on submit, but right now, I'd rather have users have an instant response than wait 6s for it to load, even if it does mean a bit more code and designing around edge cases.

Will continue to work on this and see how this solution plays out.