refactor: Change rembg function to async/await instead of .then
The problem with using .then chaining was that it would log data that wasn't fully resolved. This affected console.logs, and also when we stored the final image url into the array, it would be a Promise object.

The easiest way to fix this was to refactor using async/await and then await for all the files to be resolved, then store it or log it.

I pass this the final image urls down to the form component via props so we can run the handleSubmit all at once along with the form data.

Now, we can store all the image urls and delete them at the end once the user submits.