checkpoint: Looks Builder — items appear when a block is selected.
When the user clicks on a category, it conditionally renders the items from the wardrobe or faces from our models.

This was achieved by setting our data we want to be rendered (tops, bottoms, accessories, shoes, faces) to be under one state -> "items"

On click, we have a function that does two things:
1. Filters the data and stores it in the "items" state
2. Sets the active block to be whatever category was selected
  -> This handles our rendering and lets the app know which one the user selected. We separate these two because "items" doesn't know what is inside, it's just a place to store our data.

Once the data is filtered, and the active block is set, only then is the data mapped over inside of our modal which will display all our items or faces for our user to choose from.

Next Steps:
- When a user selects an item from the modal, it should replace the image inside of the blocks and store that id in state somewhere to be submitted to a database later on.

Resources
- Multiple onClick functions: https://stackoverflow.com/a/55296932
- Multiple conditions inside of if statement: https://stackoverflow.com/a/8710458
  -> I had an issue where doing category === "tops" || "bottoms" || etc.. wasn't working, so I had to write each condition individually as pairs.