checkpoint: Looks Builder — stores id of item when selected
When a user selects an item from the modal, it stores the id of the item in state.

The goal of this is to store that id in our database when they save a look, and then query for the objects and their image urls when we want to render it.

Next step is figuring out the query flow with Prisma.