checkpoint: Split up conditional rendering for Looks Builder, prepare for Saved Looks integration.
In the previous build, we used a ternary conditional to say: if the activeBlock was anything other than top/bottom/accessory/shoes etc., then it MUST be the model's faces.

This caused problems because when we started to use the same modal to render the Saved Looks, the faces would also be there because it falls under the "ELSE" condition.

We had to stop using a ternary and use the logical operator && to explicitly isolate both scenarios: either the user has activeBlock for wardrobe pieces, or faces.

Now when the user clicks the "View Saved" button and opens the modal, it's empty. We can now conditionally render the saved looks in this space.

Other Changes
- Updated to Next.js 12.3, next/image still buggy.
  - Created a page under campaignId called "test", use this as a testing environment.
- Created a Wardrobe Component and tried to pass down all the state as props. Everything console logs correctly, and state is passed down, but for some reason can't get it to render.

Resources
- Conditional Rendering — Inline If with Logical && Operator: https://reactjs.org/docs/conditional-rendering.html#inline-if-with-logical--operator