style/refactor: Switch item by index rather than ID, Add vertical scrolling to Looks Builder acccesories.
Instead of switching the items based on ID which has to use a findIndex and iterate through the array, we now pass the index of the item selected instead which saves us a step and also allows users to use multiples of the same items.

Without this change, if the user had two of the same items and swapped out the item, it would change the first item that matches the ID rather than the index of the block that was actually selected.

Previously, accessories would overflow past the available space on the page. overflow-y has been applied with a max-height to prevent this behaviour.

Next Steps:
- Add deletion to wardrobe
- Add "remove item" from Looks builder
- Add upload without rembg to wardrobe page.